Add directories

+++ files.go
@@ -14,7 +14,30 @@ import (
 type filesDir struct {
 	tags           string
 	id             uint64
+	dirID          id
 	renameReceiver bool
+}
+
+const (
+	negativeTag       = "_"
+	contentTag        = "@"
+	renameReceiverTag = "@@"
+)
+
+func (f filesDir) getDirectoryItem(dir string) (*item, error) {
+	if len(dir) == 0 {
+		return nil, nil
+	}
+	rows, err := f.listFiles(dir)
+	if err != nil {
+		return nil, err
+	}
+	var result item
+	if !rows.Next() {
+		return nil, syscall.ENOENT
+	}
+	db.ScanRows(rows, &result)
+	return &result, nil
 }
 
 func (f filesDir) Attr(ctx context.Context, attr *fuse.Attr) error {
@@ -27,21 +50,33 @@ func (f filesDir) Attr(ctx context.Context, attr *fuse.Attr) error {
 }
 
 func (f filesDir) listFiles(name string) (*sql.Rows, error) {
-	activeTagNames := f.getTags()
-	query := make([]string, len(activeTagNames))
-	params := make([]interface{}, len(activeTagNames)+1)
-	params[0] = []itemType{file, dir}
+	activeTagNames := f.getAllTags()
+	tagFilter := make([]string, 0, len(activeTagNames))
+	params := make([]interface{}, 0, len(activeTagNames))
+	negative := false
 	for i := range activeTagNames {
-		query[i] = "? IN tags"
-		params[i+1] = activeTagNames[i]
+		if negative {
+			negative = false
+			tagFilter = append(tagFilter, "? NOT IN tags")
+			params = append(params, activeTagNames[i])
+		} else {
+			if activeTagNames[i] == negativeTag {
+				negative = true
+			} else {
+				tagFilter = append(tagFilter, "? IN tags")
+				params = append(params, activeTagNames[i])
+			}
+		}
 	}
 	if name != "" {
 		params = append(params, name)
-		query = append(query, "i.name = ?")
+		tagFilter = append(tagFilter, "i.name = ?")
 	}
-	rows, err := db.Raw("WITH tags AS (SELECT name FROM item_tags LEFT JOIN items ON id = other_id WHERE item_id = i.id) "+
-		"SELECT * FROM items i WHERE i.type IN (?) AND "+
-		strings.Join(query, " AND "), params...).Rows()
+	tagFilter = append(tagFilter, "i.parent_id = ?", "i.type IN (?)")
+	params = append(params, f.dirID, []itemType{file, dir})
+	query := "WITH tags AS (SELECT name FROM item_tags LEFT JOIN items ON id = other_id WHERE item_id = i.id) " +
+		"SELECT * FROM items i WHERE " + strings.Join(tagFilter, " AND ")
+	rows, err := db.Raw(query, params...).Rows()
 	if err != nil {
 		return nil, err
 	}
@@ -62,10 +97,28 @@ func (f filesDir) findFile(name string) (*item, error) {
 	return &i, nil
 }
 
-func (f filesDir) getTags() []string {
+func (f filesDir) getAllTags() []string {
 	result := strings.Split(f.tags, string(os.PathSeparator))
-	if result[len(result)-1] == "@" || result[len(result)-1] == "@@" {
+	if result[len(result)-1] == contentTag || result[len(result)-1] == renameReceiverTag {
 		result = result[:len(result)-1]
+	}
+	return result
+}
+
+func (f filesDir) getTags() []string {
+	allTags := f.getAllTags()
+	var result []string
+	skipTag := false
+	for _, tag := range allTags {
+		if skipTag {
+			skipTag = false
+		} else {
+			if tag == negativeTag {
+				skipTag = true
+			} else {
+				result = append(result, tag)
+			}
+		}
 	}
 	return result
 }
@@ -98,11 +151,9 @@ func (f filesDir) ReadDirAll(ctx context.Context) ([]fuse.Dirent, error) {
 
 func (f filesDir) Create(ctx context.Context, req *fuse.CreateRequest, resp *fuse.CreateResponse) (node fs.Node, handle fs.Handle, err error) {
 	activeTagNames := f.getTags()
-	var tags = make([]item, len(activeTagNames))
-	for i := range activeTagNames {
-		db.First(&tags[i], "name = ?", activeTagNames[i])
-	}
-	var newItem = item{Name: req.Name, Type: file}
+	var tags []item
+	db.Find(&tags, "name IN (?)", activeTagNames)
+	var newItem = item{Name: req.Name, Type: file, ParentID: f.dirID}
 	db.Create(&newItem).Association("Items").Append(tags)
 	c := content{itype: file, id: uint64(newItem.ID)}
 	path, err := c.filePath()
@@ -127,6 +178,9 @@ func (f filesDir) Lookup(ctx context.Context, name string) (fs.Node, error) {
 	if i == nil {
 		return nil, syscall.ENOENT
 	}
+	if i.Type == dir {
+		return filesDir{tags: f.tags, dirID: i.ID, id: f.id + 1}, nil
+	}
 	return content{id: uint64(i.ID), itype: i.Type}, nil
 }
 
@@ -139,12 +193,18 @@ func (f filesDir) Remove(ctx context.Context, req *fuse.RemoveRequest) error {
 		return syscall.ENOENT
 	}
 	if i.ID != 0 {
-		path, err := filePath(uint64(i.ID))
-		if err != nil {
-			return err
-		}
-		if err := os.Remove(path); err != nil {
-			return err
+		if i.Type == file {
+			path, err := filePath(uint64(i.ID))
+			if err != nil {
+				return err
+			}
+			if err := os.Remove(path); err != nil {
+				return err
+			}
+		} else {
+			if !db.Find(&item{}, "parent_id = ?", i.ID).RecordNotFound() {
+				return syscall.ENOTEMPTY
+			}
 		}
 		db.Delete(&i)
 		return nil
@@ -164,17 +224,38 @@ func (f filesDir) Rename(ctx context.Context, req *fuse.RenameRequest, newDir fs
 	if srcItem == nil {
 		return syscall.ENOENT
 	}
-	tags := tagsItems(target.getTags())
+	tagsNames := target.getTags()
+	tags := tagsItems(tagsNames)
 	from, err := filePath(uint64(srcItem.ID))
 	if err != nil {
 		return err
 	}
 	srcItem.Name = req.NewName
-	db.Save(&srcItem).Association("Items").Clear().Append(tags)
+	db.Save(&srcItem).Association("Items").Replace(tags)
 	to, err := filePath(uint64(srcItem.ID))
 	if err != nil {
 		return err
 	}
 	os.Rename(from, to)
 	return nil
+}
+
+func (f filesDir) Mkdir(ctx context.Context, req *fuse.MkdirRequest) (fs.Node, error) {
+	tagsNames := f.getTags()
+	var parentID id = 0
+	var parentDir item
+	if !db.First(&parentDir, "id = ?", f.dirID).RecordNotFound() {
+		parentID = parentDir.ID
+	}
+	result := filesDir{f.tags, f.id + 1, 0, false}
+	newDir := item{0, req.Name, dir, parentID, nil}
+	var tags []item
+	if db.Find(&tags, "name in (?)", tagsNames).RecordNotFound() {
+		return nil, syscall.ENOENT
+	}
+	if err := db.Create(&newDir).Association("Items").Replace(&tags).Error; err != nil {
+		return nil, syscall.EINVAL
+	}
+	result.dirID = newDir.ID
+	return result, nil
 }
